[feature] 대화 데이터 여러 개 처리하기
