+++ After Christmas/Assets/01. Scenes/HNY/DialogueTest.unity
@@ -1353,6 +1353,7 @@ GameObject:
   - component: {fileID: 1008061621}
   - component: {fileID: 1008061623}
   - component: {fileID: 1008061624}
+  - component: {fileID: 1008061625}
   m_Layer: 0
   m_Name: Capsule
   m_TagString: Player
@@ -1531,6 +1532,36 @@ SphereCollider:
   serializedVersion: 3
   m_Radius: 3
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &1008061625
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1008061616}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 62899f850307741f2a39c98a8b639597, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_Actions: {fileID: -944628639613478452, guid: a3139f9eb3be87b46bb47c3bbe8863bd, type: 3}
+  m_NotificationBehavior: 0
+  m_UIInputModule: {fileID: 0}
+  m_DeviceLostEvent:
+    m_PersistentCalls:
+      m_Calls: []
+  m_DeviceRegainedEvent:
+    m_PersistentCalls:
+      m_Calls: []
+  m_ControlsChangedEvent:
+    m_PersistentCalls:
+      m_Calls: []
+  m_ActionEvents: []
+  m_NeverAutoSwitchControlSchemes: 0
+  m_DefaultControlScheme: PC
+  m_DefaultActionMap: Player
+  m_SplitScreenIndex: -1
+  m_Camera: {fileID: 0}
 --- !u!1 &1579639092
 GameObject:
   m_ObjectHideFlags: 0
@@ -1767,9 +1798,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7f4ae4a892db02a45b7b5bacb4b9dfc1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dialogueData: {fileID: 11400000, guid: cac610d00ea13a84e95fe6737763aebd, type: 2}
-  characterData: {fileID: 11400000, guid: 755763d612b95744e81c60d91ed6c176, type: 2}
-  speechBubble: {fileID: 1687603966}
+  dialogueSequence: []
   playerTag: Player
 --- !u!1001 &1653929210
 PrefabInstance:
@@ -1896,9 +1925,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 7f4ae4a892db02a45b7b5bacb4b9dfc1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  dialogueData: {fileID: 11400000, guid: b2cfde31110b1c447b2db7fee8d35c90, type: 2}
-  characterData: {fileID: 11400000, guid: 4c929a4e8c153934295b1fc9b95ebd29, type: 2}
-  speechBubble: {fileID: 742604769}
+  dialogueSequence: []
   playerTag: Player
 --- !u!1 &1683113254
 GameObject:
@@ -1961,7 +1988,7 @@ Transform:
   m_GameObject: {fileID: 1687603966}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: -0.92387956, z: 0, w: 0.38268343}
-  m_LocalPosition: {x: -2, y: 4.2666664, z: 2}
+  m_LocalPosition: {x: -2.02, y: 4.2666664, z: 1.98}
   m_LocalScale: {x: 0.33333343, y: 0.3333333, z: 0.33333343}
   m_ConstrainProportionsScale: 0
   m_Children: []
