+++ After Christmas/Assets/02. Scripts/Interface/IInteractable.cs
@@ -4,3 +4,4 @@ public interface IInteractable
 {
     void Interact();
 }
+
